refactor: update src folder architecture
